Показываем кнопку today tasks только если задачи на сегодня есть
