+++ db.json
@@ -97,20 +97,12 @@
       "date": "2023-02-27T21:00:00.000Z"
     },
     {
-      "id": "6fa5be43-74bf-44e1-bff7-090733b68300",
-      "title": "new task",
-      "description": "dafsfgsrfgrsg",
-      "colorBadge": "green.500",
+      "id": "3102d702-f946-4e9a-a885-6ece1e3f3039",
+      "title": "Kaef",
+      "description": "Big",
+      "colorBadge": "blue.500",
       "opened": true,
-      "date": "2023-01-23T11:42:56.505Z"
-    },
-    {
-      "id": "b9ec2f92-e1de-4040-bb95-30f32222407a",
-      "title": "Title ",
-      "description": "Description ",
-      "colorBadge": "gray.500",
-      "opened": true,
-      "date": "2023-01-23T21:00:00.000Z"
+      "date": "2023-01-22T21:00:00.000Z"
     }
   ]
 }
