+++ src/utils/todos-utils.ts
@@ -16,3 +16,34 @@ export const formatTodosForAccordion = (todos: ITodo[]) => {
 
   return newTodos;
 };
+
+export const converDataToTodayOrYestoday = (date: string) => {
+  const timeStamp = new Date().getTime();
+  const singleDay = 24 * 60 * 60 * 1000;
+
+  const yesterday = format(new Date(timeStamp - singleDay), "MM/dd/yyyy");
+  const tomorrow = format(new Date(timeStamp + singleDay), "MM/dd/yyyy");
+  const today = format(new Date(), "MM/dd/yyyy");
+
+  if (yesterday == date) {
+    return "Yesterday";
+  }
+
+  if (tomorrow == date) {
+    return "Tomorrow";
+  }
+
+  if (today == date) {
+    return "Today";
+  }
+
+  return date;
+};
+
+export const returnTodayDate = () => {
+  return format(new Date(), "MM/dd/yyyy");
+};
+
+export const isListHasTodayTask = (list: { [key: string]: ITodo[] }) => {
+  return list[returnTodayDate()];
+};
